update: Base.java Result.java

+++ src/main/java/com/group6/pojo/Bill.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -16,10 +17,10 @@ public class Bill {
     private int days;
     private int building;
     private int dormitory;
-    private BigDecimal electricity_usage;
-    private BigDecimal electricity_cost;
-    private BigDecimal water_usage;
-    private BigDecimal water_cost;
-    private BigDecimal total_cost;
+    private double electricity_usage;
+    private double electricity_cost;
+    private double water_usage;
+    private double water_cost;
+    private double total_cost;
 
 }
